Add utils.getComponentTree Refactor the Rehype code into `utils.getComponentTree({ components, html, settings })`.
